+++ platforms/core-execution/execution/src/main/java/org/gradle/internal/execution/steps/PreCreateOutputParentsStep.java
@@ -25,12 +25,12 @@ import java.io.File;
 
 import static org.gradle.util.internal.GFileUtils.mkdirs;
 
-public class CreateOutputsStep<C extends ChangingOutputsContext, R extends Result> implements Step<C, R> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CreateOutputsStep.class);
+public class PreCreateOutputParentsStep<C extends ChangingOutputsContext, R extends Result> implements Step<C, R> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(PreCreateOutputParentsStep.class);
 
     private final Step<? super C, ? extends R> delegate;
 
-    public CreateOutputsStep(Step<? super C, ? extends R> delegate) {
+    public PreCreateOutputParentsStep(Step<? super C, ? extends R> delegate) {
         this.delegate = delegate;
     }
