+++ platforms/core-execution/execution/src/test/groovy/org/gradle/internal/execution/steps/PreCreateOutputParentsStepTest.groovy
@@ -20,8 +20,8 @@ import org.gradle.api.internal.file.TestFiles
 import org.gradle.internal.execution.UnitOfWork
 import org.gradle.internal.file.TreeType
 
-class CreateOutputsStepTest extends StepSpec<ChangingOutputsContext> {
-    def step = new CreateOutputsStep<>(delegate)
+class PreCreateOutputParentsStepTest extends StepSpec<ChangingOutputsContext> {
+    def step = new PreCreateOutputParentsStep<>(delegate)
 
     def "outputs are created"() {
         given:
